Move bill run generator logic to 'generator'
We don't want all the generation code sat in the controller. So, initially a 'service' seemed the best place for it. But we had already come to the conclusion we are comfortable with this code not having direct unit test coverage. We are also uncomfortable with code designed to support testing of the app sitting with the app code proper.

So, then we hit upon the idea of moving the logic to the `/test` folder to make these distinctions clear. This should also help us with external tools like SonarCloud.

With this change we simplify the test controller. It can still create the bill run; all that functionality is already in the `CreateBillRunService` so there seemed no need to duplicate it. But then we make use of our new generator class.
