+++ test/support/generators/index.js
@@ -1,0 +1,7 @@
+'use strict'
+
+const BillRunGenerator = require('./bill_run.generator')
+
+module.exports = {
+  BillRunGenerator
+}
